Remove explicit Int declaration of peakfinder_σ_scaled

+++ src/sipm_simple_calibration.jl
@@ -46,7 +46,7 @@ function sipm_simple_calibration(pe_uncal::Vector{<:Real};
         bin_width_cut_scaled = bin_width_cut * bin_width_scale
         @debug "Using bin width: $(bin_width_cut_scaled)"
         h_uncal_cut = fit(Histogram, pe_uncal, bin_width_cut_min:bin_width_cut_scaled:initial_max_amp)
-        peakfinder_σ_scaled::Int = if peakfinder_σ <= 0.0
+        peakfinder_σ_scaled = if peakfinder_σ <= 0.0
             round(Int, 2*(cuts_1pe.high - cuts_1pe.max) / bin_width_cut_scaled / (2 * sqrt(2 * log(2))) )
         else
             rount(Int, peakfinder_σ)  
@@ -75,7 +75,7 @@ function sipm_simple_calibration(pe_uncal::Vector{<:Real};
         @debug "Using bin width: $(bin_width_cut_scaled)"
         
         h_uncal_cut = fit(Histogram, pe_uncal, bin_width_cut_min:bin_width_cut_scaled:initial_max_amp)
-        peakfinder_σ_scaled::Int = if peakfinder_σ <= 0.0
+        peakfinder_σ_scaled = if peakfinder_σ <= 0.0
             round(Int, 2*(cuts_1pe.high - cuts_1pe.max) / bin_width_cut_scaled / (2 * sqrt(2 * log(2))) )
         else
             rount(Int, peakfinder_σ)
